Put aria-live region back inside cpanel.
Closes #365.
